+++ src/@dev--com.example; example userscript.user.ts
@@ -4,7 +4,7 @@
 // @version     1.0.0
 // @description 2025/03/31, 5:38:51 PM
 // @run-at      document-start
-// @grant       none
+// @grant       GM_addElement
 // @homepageURL https://github.com/ericchase-library/ts-templates-browser-userscript
 // ==/UserScript==
 
@@ -12,13 +12,15 @@
 // UserScript and script for refreshing the page when relevant files are
 // modified.
 
-import { WebPlatform_DOM_Inject_Script } from './lib/ericchase/WebPlatform_DOM_Inject_Script.js';
 import { Async_WebPlatform_DOM_ReadyState_Callback } from './lib/ericchase/WebPlatform_DOM_ReadyState_Callback.js';
 import { SERVERHOST } from './lib/server/info.js';
 
 Async_WebPlatform_DOM_ReadyState_Callback({
   async DOMContentLoaded() {
-    WebPlatform_DOM_Inject_Script(await fetch(`http://${SERVERHOST()}/com.example; example userscript.user.js`).then((response) => response.text()));
-    WebPlatform_DOM_Inject_Script(await fetch(`http://${SERVERHOST()}/lib/server/hot-reload.iife.js`).then((response) => response.text()));
+    // We are able to bypass browser CSP features through the UserScript
+    // manager's dedicated API. In the future, I'd like to automatically
+    // generate these dev scripts via build tools.
+    GM_addElement('script', { src: `http://${SERVERHOST()}/com.example; example userscript.user.js` });
+    GM_addElement('script', { src: `http://${SERVERHOST()}/lib/server/hot-reload.iife.js` });
   },
 });