+++ CHANGELOG.md
@@ -2,6 +2,12 @@ Please refer to the `ts-library` changelog for library and build tool changes:
 
 - https://github.com/ericchase-library/ts-library/blob/main/CHANGELOG.md
 
+## 2025-09-01
+
+- Some minor changes here and there
+- Updated the `@dev` UserScript to use the UserScript manager's API for injecting elements into the page
+  - With this, we can bypass the browser's CSP feature
+
 ## 2025-08-16
 
 - Updated `Processor_TypeScript_UserScript_Bundler` processor to handle `import.meta.url`